updated tests.

+++ calculator/src/main/java/com/calculator/entities/Multiply.java
@@ -4,7 +4,7 @@ public class Multiply {
 
     private int total = 0;
 
-    public int totalMultiplier(int mult1, int mult2){
+    public double totalMultiplier(int mult1, int mult2){
        total = mult1 * mult2;
        return total;
     }
