+++ calculator/src/main/java/com/calculator/entities/Subtract.java
@@ -4,7 +4,7 @@ public class Subtract {
 
     private int total = 0;
 
-    public int totalSubtraction(int sub1, int sub2){
+    public double totalSubtraction(int sub1, int sub2){
        total = sub1 - sub2;
        return total;
     }
