+++ calculator/src/main/java/com/calculator/entities/Sum.java
@@ -4,7 +4,7 @@ public class Sum {
 
     private int total = 0;
 
-    public int totalSum(int sum1, int sum2){
+    public double totalSum(int sum1, int sum2){
        total = sum1 + sum2;
        return total;
     }
